paginate companies, updated imageservice

+++ webroot/app/Services/ImageService.php
@@ -26,15 +26,29 @@ class ImageService
             return;
         }
 
-        Storage::delete($resource->logo->url);
-
         $url = $request->file('image')->store(self::IMAGES_PATH);
-        $resource->logo->update(['url' => $url]);
+
+        if (isset($resource->logo)) {
+            $this->deleteFile($resource->logo->url);
+            $resource->logo->update(['url' => $url]);
+        } else {
+            $image = new Image(['url' => $url]);
+            $resource->images()->save($image);
+        }
     }
 
-    public function deleteImage(Image $image): void
+    public function deleteImage(Image $image = null): void
     {
-        Storage::delete($image->url);
-        $image->delete();
+        if ($image) {
+            Storage::delete($image->url);
+            $image->delete();
+        }
+    }
+
+    private function deleteFile(string $url = null): void
+    {
+        if (Storage::exists($url)) {
+            Storage::delete($url);
+        }
     }
 }
